Make Keys generic over "instant" type to use
So far our Keys type has assumed that what is being passed in as
representing the current time is of type std::time::Instant. However, in
certain environment it may not be possible to acquire an instance of
such a type.
In such cases it can be important for the logic to be generic over the
type of instant being worked with. With this change we make that happen.