Use HashMap::retain() in Keys::tick() logic
The logic in Keys::tick() basically iterates over all currently pressed
keys and removes those that have been released or whose auto-repeated
has been disabled. Because of the usage of HashMap::iter_mut(), it can't
perform the removal as part of the iteration. Hence, it goes to some
lengths to make it happen after the fact, but with minimal effort
involved.
As it turns out, the HashMap::retain() method can help us here, as it
basically combines the iteration aspect with that of removal in an
easy-to-use package. Switch over to using it to simplify the code.